requirements.txt
pillow==10.4.0
streamlit>=1.37.0
